mqtt-otel-bridge/requirements.txt
paho-mqtt==1.6.1
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
grpcio==1.62.1
influxdb-client3==0.5.0
